refactor: inquiry modal tests
